+++ system/lib/standalone_wasm.c
@@ -8,7 +8,6 @@
 #include <assert.h>
 #include <emscripten.h>
 #include <errno.h>
-#include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 #include <time.h>
@@ -140,18 +139,6 @@ int emscripten_resize_heap(size_t size) {
   }
 #endif
   return 0;
-}
-
-void emscripten_console_log(const char *utf8String) {
-  puts(utf8String);
-}
-
-void emscripten_console_warn(const char *utf8String) {
-  fprintf(stderr, "%s\n", utf8String);
-}
-
-void emscripten_console_error(const char *utf8String) {
-  emscripten_console_warn(utf8String);
 }
 
 // C++ ABI
