+++ tests/other/metadce/hello_libcxx_O2.sent
@@ -2,6 +2,7 @@ __cxa_atexit
 __map_file
 __sys_munmap
 abort
+emscripten_console_error
 emscripten_get_sbrk_ptr
 emscripten_memcpy_big
 emscripten_resize_heap
