+++ tests/other/metadce/hello_libcxx_O2_fexceptions.imports
@@ -11,6 +11,7 @@ __map_file
 __resumeException
 __sys_munmap
 abort
+emscripten_console_error
 emscripten_memcpy_big
 emscripten_resize_heap
 environ_get
