+++ tests/other/metadce/hello_libcxx_O2_fexceptions.sent
@@ -12,6 +12,7 @@ __map_file
 __resumeException
 __sys_munmap
 abort
+emscripten_console_error
 emscripten_get_sbrk_ptr
 emscripten_memcpy_big
 emscripten_resize_heap
