+++ tests/other/metadce/hello_libcxx_O2_fexceptions_DEMANGLE_SUPPORT.funcs
@@ -423,7 +423,6 @@ $__multf3
 $__multi3
 $__munmap
 $__newlocale
-$__overflow
 $__shgetc
 $__shlim
 $__stdio_close
@@ -507,7 +506,6 @@ $fmt_fp
 $fmt_o
 $fmt_u
 $fmt_x
-$fputc
 $freelocale
 $frexp
 $fwrite
@@ -1232,7 +1230,6 @@ $unsigned\20long\20std::__2::\28anonymous\20namespace\29::countof<char>\28char\2
 $unsigned\20long\20std::__2::\28anonymous\20namespace\29::countof<wchar_t>\28wchar_t\20const*\2c\20wchar_t\20const*\29
 $unsigned\20short\20std::__2::__num_get_unsigned_integral<unsigned\20short>\28char\20const*\2c\20char\20const*\2c\20unsigned\20int&\2c\20int\29
 $vasprintf
-$vfprintf
 $vfscanf
 $virtual\20thunk\20to\20std::__2::basic_istream<char\2c\20std::__2::char_traits<char>\20>::~basic_istream\28\29
 $virtual\20thunk\20to\20std::__2::basic_istream<char\2c\20std::__2::char_traits<char>\20>::~basic_istream\28\29.1
